requirements-app.txt
fastapi
uvicorn
httpx
pydantic
mlflow
langfuse
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
